Update existing deobfs to use new format
