+++ src/vipyr_deobf/cli.py
@@ -1,115 +1,75 @@
 # PYTHON_ARGCOMPLETE_OK
 import argparse
-import logging
 import importlib.util
-from typing import Callable, TypeVar
+import logging
 
-from .utils import Color, setup_logging
-from .deobfuscators.blankobf2 import deobf_blankobf2, format_blankobf2
-from .deobfuscators.fct import deobf_fct, format_fct
-from .deobfuscators.hyperion import deobf_hyperion, format_hyperion
-from .deobfuscators.lzmaspam import deobf_lzma_b64, format_lzma_b64
-from .deobfuscators.pyobfuscate import deobf_pyobfuscate, format_pyobfuscate
-from .deobfuscators.vare import deobf_vare, format_vare
-from .exceptions import DeobfuscationFailError
-from .scanners.blankobf2_scan import scan_blankobf2
-from .scanners.fct_scan import scan_fct
-from .scanners.hyperion_scan import scan_hyperion
-from .scanners.lzmaspam_scan import scan_lzma
-from .scanners.pyobfuscate_scan import scan_pyobfuscate
-from .scanners.vare_scan import scan_vare
-
-R = TypeVar('R')
-
-supported_obfuscators: dict[str, tuple[Callable[[str], R], Callable[[R], str]]] = {
-    'hyperion': (deobf_hyperion, format_hyperion),
-    'lzmaspam': (deobf_lzma_b64, format_lzma_b64),
-    'vare': (deobf_vare, format_vare),
-    'fct': (deobf_fct, format_fct),
-    'blankobf2': (deobf_blankobf2, format_blankobf2),
-    'pyobfuscate': (deobf_pyobfuscate, format_pyobfuscate),
-}
-
-scanners: dict[str, Callable[[str], bool]] = {
-    'hyperion': scan_hyperion,
-    'lzmaspam': scan_lzma,
-    'vare': scan_vare,
-    'fct': scan_fct,
-    'blankobf2': scan_blankobf2,
-    'pyobfuscate': scan_pyobfuscate,
-}
-
-alias_dict: dict[str, str] = {
-    'vore': 'vare',
-    'hyperd': 'hyperion',
-    'fct_obfuscate': 'fct',
-    'not_pyobfuscate': 'fct',
-    'blankobfv2': 'blankobf2',
-}
+from vipyr_deobf.deobf_base import iter_deobfs, load_all_deobfs, load_deobfs
+from vipyr_deobf.exceptions import DeobfuscationFailError
+from vipyr_deobf.utils import Color, setup_logging
 
 
-
-def run_deobf(code: str, deobf_type: str) -> str:
-    deobf_func, format_func = supported_obfuscators[deobf_type]
-    results = deobf_func(code)
-    return format_func(*results) if isinstance(results, tuple) else format_func(results)
-
-
-def run():
+def get_parser():
     parser = argparse.ArgumentParser(
         prog='Vipyr Deobfuscator',
         description='Deobfuscates obfuscated scripts',
     )
     parser.add_argument('path', help='path to obfuscated file')
     parser.add_argument('-t', '--type', default='auto', type=str,
-                        choices=list(supported_obfuscators.keys())+['auto'],
-                        help='type of obfuscation used (defaults to auto)')
+                        help='type of obfuscation used, see README for options (defaults to auto)')
     parser.add_argument('-o', '--output', help='file to output deobf result to, defaults to stdout')
-    parser.add_argument('-d', '--debug', action='store_true', help='display debug logs (defaults to false)')
-    parser.add_argument('-s', '--soft', action='store_true', help='display expected warnings (defaults to false)')
+    parser.add_argument('-s', '--skip-scan', action='store_true', help='skip scanning phase to identify schema')
+    parser.add_argument('-d', '--debug', action='store_true', help='display debug logs')
+    parser.add_argument('--show-expected', action='store_true', help='display expected warnings')
+    return parser
+
+
+def run():
+    parser = get_parser()
     if importlib.util.find_spec('argcomplete') is not None:
         import argcomplete
         argcomplete.autocomplete(parser)
+
     args = parser.parse_args()
 
     logger = logging.getLogger('deobf')
     setup_logging(args)
+    logger.info('Logging setup finished')
 
+    logger.info(f'Opening file at {args.path}')
     try:
         with open(args.path, 'r') as file:
             data = file.read()
     except FileNotFoundError:
         logger.error(f'{args.path} is not a valid path.')
         return
+    logger.info('Data successfully read from file')
 
-    schemas = []
+    logger.info('Loading deobfuscators...')
     if args.type == 'auto':
-        logger.info('Scanning file to identify schema')
-        for schema, scanner in scanners.items():
-            if scanner(data):
-                schemas.append(schema)
-        if not schemas:
-            logger.error('Could not identify obfuscation schema')
-            return
-        logger.info(f'Schemas matched: {", ".join(schemas)}')
+        load_all_deobfs()
     else:
-        deobf_type = args.type.replace('-', '_')
-        deobf_type = alias_dict.get(deobf_type, deobf_type)
-        if deobf_type not in supported_obfuscators:
-            logger.error(
-                f'Unsupported obfuscation schema.\n'
-                f'Supported obfuscation schemes include:\n'
-                f'{", ".join(supported_obfuscators)}'
-            )
-            return
-        schemas.append(deobf_type)
+        load_deobfs(args.type)
 
-    for schema in schemas:
+    if args.skip_scan:
+        deobfs = [*iter_deobfs()]
+    else:
+        logger.info('Running scanners...')
+        deobfs = []
+        for deobf in iter_deobfs():
+            logger.info(f'Scanning with schema {deobf.name}v{deobf.version}')
+            if deobf.scan(data):
+                logger.info('Scan succeeded, adding to schema list')
+                deobfs.append(deobf)
+            else:
+                logger.info('Scan failed, skipping')
+    logger.info(f'Schema list: {", ".join([deobf.name for deobf in deobfs])}')
+
+    for deobf in deobfs:
         try:
-            logger.info(f'Running deobf of {args.path} with schema <{schema}>')
-            output = run_deobf(data, schema)
+            logger.info(f'Running deobf of {args.path} with schema {deobf.name}')
+            output = deobf.deobf(data)
         except DeobfuscationFailError as exc:
-            logger.exception(f'Deobfuscation of {args.path} with schema <{schema}> failed:')
+            logger.exception(f'Deobfuscation of {args.path} with schema {deobf.name} failed:')
             for var, data in exc.env_vars.items():
                 print(f'{Color.bold_red}{var}{Color.clear}', data, sep='\n', end='\n\n')
         else:
@@ -119,7 +79,4 @@ def run():
                 logger.info(f'Writing results of deobf to file {args.output}')
                 with open(args.output, 'w') as file:
                     file.write(output)
-
-
-if __name__ == '__main__':
-    run()
+            break
