--- src/vipyr_deobf/scanners/pyobfuscate_scan.py
@@ -1,9 +1,0 @@
-import ast
-
-from ..deobfuscators.pyobfuscate import PayloadExtractor
-
-
-def scan_pyobfuscate(code: str):
-    extractor = PayloadExtractor()
-    extractor.visit(ast.parse(code))
-    return bool(extractor.obf_type)
