+++ tests/blankobf/v2/test_blankobf_v2.py
@@ -1,6 +1,6 @@
 import ast
 
-from vipyr_deobf.deobfuscators import blankobf2
+from vipyr_deobf.deobfuscators.BlankObf.blankobf_v2 import blankobf_v2_deobf
 
 BLANK_OBF_HEADER = ':: You managed to break through BlankOBF v2; Give yourself a pat on your back! ::'
 
@@ -8,7 +8,7 @@ BLANK_OBF_HEADER = ':: You managed to break through BlankOBF v2; Give yourself a
 def test_deobf_hello_world():
     with open('tests/blankobf/v2/sample_array.obf', 'r') as file:
         obf = file.read()
-    status, res = blankobf2.deobf_blankobf2(obf)
+    status, res = blankobf_v2_deobf.deobf(obf)
     assert status
     with open('tests/blankobf/v2/sample_array.exp', 'r') as file:
         exp = file.read()
